Move files and update enemy steering behaviour #182 #179
Enemy now uses a combination of flock, wander, and path following to steer. This looks a more realistic and works using only the Rigidbody. Any forces that are applied to the rigidbody will work just fine. The "PathFollower" game object is now integrated in the PathFollow steering behaviour to reduce the amount of objects in the scene.